+++ cmd/ixf/main.go
@@ -26,7 +26,7 @@ import (
 
 const defaultCookies = "/tmp/ixunfei_profile_explorer_cookies.json"
 
-var version = "3.6.1"
+var version = "3.6.2"
 
 var skillNames = []string{
 	"using-ixf-toolbox",